add data.feature mocks
