+++ src/data.test.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import { initialize } from './index';
+import { initialize } from "./index";
 
-test('data object is mocked', () => {
+test("data object is mocked", () => {
   initialize();
   expect(new google.maps.Data(null)).toBeTruthy();
   const map = new google.maps.Map(null);
